Restructure all code
