+++ .gitignore
@@ -107,3 +107,4 @@ kaydara_env
 .vscode
 
 dados
+req
