+++ requirements.txt
@@ -20,13 +20,13 @@ Jinja2==2.10
 MarkupSafe==1.0
 newsapi-python==0.0.2
 oauth2client==4.1.2
+psycopg2-binary==2.7.4
 pyasn1==0.4.2
 pyasn1-modules==0.2.1
 pycparser==2.18
 pymessenger==0.0.7.0
 pyOpenSSL==17.5.0
 python-dateutil==2.7.2
-psycopg2-binary==2.7.4
 requests==2.13.0
 requests-toolbelt==0.8.0
 rsa==3.4.2
