+++ src/APIs/NewsAPI.py
@@ -1,5 +1,4 @@
 import os, json, re
-from sys import argv
 from flask import url_for
 from newsapi import NewsApiClient
 
@@ -29,22 +28,3 @@ def isImageAvailable(url):
     if img == None:
         return False
     return True
-
-# Generate news articles to send.
-def generateNewsPosts(data):
-    posts = []
-    for element in data:
-        post = {
-            "title": element['title'],
-            "image_url": element['urlToImage'],
-            "subtitle": element['description'],
-            "buttons": [
-                {
-                    'type': 'web_url',
-                    'url': element['url'],
-                    'title': 'Read more'
-                }
-            ]
-        }
-        posts.append(post)
-    return posts
