+++ src/NLP/WeatherNLP.py
@@ -1,15 +1,17 @@
-from Utils import IBMWatsonAPI
-from Utils import FacebookAPI
-from Utils import OpenWeatherMapAPI as WeatherAPI
-from Utils.db import example
+from src.APIs import IBMWatsonAPI
+from src.APIs import FacebookAPI
+from src.APIs import OpenWeatherMapAPI as WeatherAPI
+from src.MsgBuilder import WeatherMB
+from src.Models import Client
 import json
 
 
-def process_message(msg, cli):
-    (intents, entities, newContext, output) = IBMWatsonAPI.send_message(msg, cli.context)
+def process_message(results, cli):
+    print('WEATHER REQUEST')
+    (intents, entities, newContext, output) = results
 
     json_newContext = json.dumps(newContext, indent=2)
-    example.update_client_context(cli.id, json_newContext)
+    Client.update_client_context(cli.id, json_newContext)
 
     for m in output:
         FacebookAPI.send_message(cli.id, m)
@@ -25,16 +27,16 @@ def process_message(msg, cli):
 def simple_forecast(context, cli, hour=False):
     prevision = WeatherAPI.getForecastCity(context['location'])
     if hour:
-        icon, title, subtitle = WeatherAPI.generateSimpleForecastDay(prevision, context['date'], context['time'])
+        icon, title, subtitle = WeatherMB.generateSimpleForecastDay(prevision, context['date'], context['time'])
         FacebookAPI.send_picture(cli.id, icon, title, subtitle)
     else:
-        elements = WeatherAPI.generateForecastDay(prevision, context['date'])
+        elements = WeatherMB.generateForecastDay(prevision, context['date'], True)
         FacebookAPI.send_list(cli.id, elements)
-    example.update_client_context(cli.id, None)
+    Client.update_client_context(cli.id, None)
 
 
 def next_days_forecast(context, cli):
     prevision = WeatherAPI.getForecastCity(context['location'])
-    elements = WeatherAPI.generateForecastPosts(prevision)
+    elements = WeatherMB.generateForecastPosts(prevision)
     FacebookAPI.send_list(cli.id, elements)
-    example.update_client_context(cli.id, None)
+    Client.update_client_context(cli.id, None)
